requirements.txt
# --- 핵심 프레임워크 ---
streamlit==1.41.1
langchain==0.3.17
langchain-community==0.3.15
langchain-upstage==0.7.0
langchain-openai==0.3.0
langgraph==0.2.66
langchain-core==0.3.33

# --- 벡터 DB 및 검색 ---
faiss-cpu==1.9.0.post1

# --- 의존성 충돌 해결 ---
packaging==24.2
protobuf==4.25.5
pydantic==2.10.6
pydantic-settings==2.7.1

# --- 데이터 및 유틸리티 ---
pandas>=2.2.0
python-dotenv>=1.0.1
requests>=2.32.0